requirements.txt
dgl-cu101==0.7.0
numpy==1.19.3
ogb==1.3.1
torch==1.7.1+cu101
scikit-learn==0.24.2
tqdm==4.57.0

